+++ app/models/concerns/review_images.rb
@@ -1,8 +1,8 @@
-module Images
+module ReviewImages
   extend ActiveSupport::Concern
   included do
     def display_images(height: 180, weight: 180)
-      images.includes([:blob]).each { |image| image.variant(resize_to_limit: [height, weight]).processed }
+      images.includes([:blob]).each { |image| image.variant(resize_to_fill: [height, weight]).processed }
     end
 
     private
