+++ app/views/admins/articles/_form.html.slim
@@ -28,10 +28,5 @@
         .form-group
           = f.label :address
           = f.text_field :address, class: "form-control mb-2"
-          input type="button" value="位置検索" onclick="codeAddress()"
-
-        #display
-        #map style="width: auto; height:400px"
-
         .actions.mt-4
           = f.submit
