+++ app/views/admins/articles/index.html.slim
@@ -9,3 +9,4 @@
           = link_to "記事を作成する", new_admins_article_path, class: "btn btn-primary btn-block"
         ol
           = render partial: "article", collection: @articles
+        = paginate @articles
