+++ app/views/admins/articles/new.html.slim
@@ -1,38 +1,1 @@
 = render "form", article: @article, url: admins_articles_path
-
-javascript:
-  let map
-  let geocoder
-  const display = document.getElementById('display')
-
-  function initMap(){
-    geocoder = new google.maps.Geocoder()
-
-    map = new google.maps.Map(document.getElementById('map'), {
-      center: {lat: 35.6803997, lng: 139.7690174},
-      zoom: 15,
-    });
-
-    marker = new google.maps.Marker({
-      position:  {lat: 35.6803997, lng: 139.7690174},
-      map: map
-    });
-  }
-
-  function codeAddress(){
-    let inputAddress = document.getElementById('article_address').value;
-
-    geocoder.geocode( { 'address': inputAddress}, function(results, status) {
-      if (status == 'OK') {
-        map.setCenter(results[0].geometry.location);
-        var marker = new google.maps.Marker({
-            map: map,
-            position: results[0].geometry.location
-        });
-        display.textContent = "検索結果：" + results[ 0 ].geometry.location
-      } else {
-        alert('該当する結果がありませんでした：' + status);
-      }
-    });
-  }
-script src="https://maps.googleapis.com/maps/api/js?key=#{Rails.application.credentials.googlemap[:api_key]}&callback=initMap" async defer
