+++ app/views/layouts/_breadcrumbs.html.slim
@@ -1,6 +1,6 @@
 - breadcrumbs.tap do |links|
   - if links.any?
-    .container.d-none.d-md-block.mb-4
+    .container.d-none.d-md-block
       ul.breadcrumb.bg-transparent
         - links.each do |link|
           - if link.current?
