+++ app/views/layouts/_header_navbar_list.html.slim
@@ -2,15 +2,11 @@ li class="nav-item #{nav_class}"
   = link_to gunplas_path, class: "nav-link #{nav_color}"
     i.fas.fa-search.mr-2
     | ガンプラリスト
-/ li class="nav-item #{nav_class}"
-/   a class="nav-link #{nav_color}" href="#"
-/     i.fas.fa-star.mr-2
-/     | レビューリスト
 li class="nav-item #{nav_class}"
   = link_to rankings_path, class: "nav-link #{nav_color}"
     i.fas.fa-crown.mr-2
     | ランキング
 li class="nav-item #{nav_class}"
-  a class="nav-link #{nav_color}" href="#"
+  = link_to articles_path, class: "nav-link #{nav_color}"
     i.fas.fa-rss.mr-2
     | ニュース
