+++ app/views/users/articles/_article.html.slim
@@ -1,0 +1,10 @@
+= link_to article_path(article)
+  li.media.border.rounded.mb-2
+    .shadow-sm.mr-3 style="width: 150px;"
+      = article.image_attached?(height: 150, width: 150)
+    .media-body.my-auto
+      p
+        = article.title
+      ul.text-left.p-0
+        li.text-muted
+          = "更新日 #{article.updated_at.to_s(:datetime_jp)}"
