+++ app/views/users/articles/show.html.erb
@@ -1,13 +1,11 @@
+<% provide(:title, "記事詳細") %>
+<% breadcrumb :article, @article %>
+
 <div class="article row mx-auto">
   <div class="col-lg-8 mx-auto">
     <div class="card shadow-sm m">
       <div class="card-body m-4">
         <div class="article_head text-center mb-5">
-          <div class="d-flex justify-content-between m-4">
-            <%= link_to "編集", edit_admins_article_path(@article), class: "btn btn-success"%>
-            <%= link_to "削除", admins_article_path(@article), method: :delete, data: { confirm: "記事を削除してよろしいですか?" }, class: "btn btn-danger"%>
-            <%= link_to "戻る", admins_articles_path(@article), class: "btn btn-secondary"%>
-          </div>
           <h2 class="headingLv mb-3">
             <%= @article.title %>
           </h2>
@@ -18,7 +16,7 @@
             <%= image_tag(@article.display_image) %>
           <% end %>
         </div>
-        <div class="article_box">
+        <div class="article_box mb-4">
           <h5 class="heading">Content</h5>
           <div class="article_content mb-4">
             <%= @article.content %>
@@ -30,11 +28,13 @@
             <% end%>
           </div>
         </div>
+        <%= link_to "記事一覧へ戻る", articles_path, class: "btn btn-secondary btn-block"%>
       </div>
     </div>
   </div>
-</div>
 
+  <%= render "users/shared/sidebar" %>
+</div>
 
 <script>
   function initMap() {
